more

+++ python/sglang/srt/two_batch_overlap.py
@@ -385,7 +385,8 @@ class TboForwardBatchPreparer:
         value_a = min(tbo_split_token_index, num_token_non_padded)
         value_b = max(0, num_token_non_padded - tbo_split_token_index)
         return torch.tensor(
-            [value_a, value_b], device=global_server_args_dict["device"]
+            [value_a, value_b], device=global_server_args_dict["device"],
+            dtype=torch.int32,
         )
